+++ packages/account-node/CHANGELOG.md
@@ -1,3 +1,8 @@
+2024-10-23 (v.4.1.0)
+-------------------
+
+- Support `SessionAccount` object that can load a network connection when provided with a base64 encoded full session object
+
 2024-07-15 (v.4.0.0)
 -------------------